First pass merge by multicaller.
This replaces the merge by common caller. Instead of looking for the
least common caller of two partitions, it looks for the least caller
that is known to have contained multiple comms (based on function name).
This way it won't over-merge when a function only calls one comm event
for one task instead of many.